rename query_suggestions -> suggestions
